Added armor to Tanks
Armor decreases damage taken
Damage is now CGFloat
Tank health points now in CGFloat form
Projectile now has Ammo property rather than blastRadius and damage properties

+++ Backend/Ammo.swift
@@ -23,9 +23,9 @@ import Foundation
 class Ammo: Item {
 
 	var blastRadius: CGFloat
-	var damage: Int
+	var damage: CGFloat
 
-	init(_ name: String, price: Int, radius: CGFloat, damage: Int) {
+	init(_ name: String, price: Int, radius: CGFloat, damage: CGFloat) {
 		blastRadius = radius
 		self.damage = damage
 		super.init(name, price: price)
